requirements.txt
# requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers
accelerate
gradio
pillow
numpy
omegaconf
safetensors